moved some methods, renamed some classes
